DevelopChart/requirements2.txt
djangorestframework
PyMySQL
mysqlclient
